Add connection to Google Drive
This uses the `google_drive` gem and is getting set up the first time
the `bin/setup.rb` script is being run. The user will be prompted to
provide the Google credentials and is then asked to select a folder via
its ID that is being used for the blog.